restore getRuleExpansionMethod
