+++ packages/javascript/bh-shared-ui/src/views/PrivilegeZones/Save/RuleForm/rule-form-utils.ts
@@ -1,22 +1,4 @@
-import {
-    SeedExpansionMethod,
-    SeedExpansionMethodAll,
-    SeedExpansionMethodChild,
-    SeedExpansionMethodNone,
-} from 'js-client-library';
-
 export const emptyFunction = () => {};
-
-export const getRuleExpansionMethod = (
-    tagId: string,
-    tagType: 'labels' | 'zones',
-    ownedId: string | undefined
-): SeedExpansionMethod => {
-    // Owned is a specific tag type that does not undergo expansion
-    if (tagId === ownedId) return SeedExpansionMethodNone;
-
-    return tagType === 'zones' ? SeedExpansionMethodAll : SeedExpansionMethodChild;
-};
 
 export const CYPHER_MUST_HAVE_RESULTS =
     'To save a rule created using Cypher, the Cypher query must produce at least one result. Please run a different query to proceed.';
